refactor(editor): extract collab session, snapshot scheduler, and sketch doc ops
Move the StrictMode-surviving session refcounting + seed one-shot tracking
to collab_session.ts, the snapshot payload/debounce/retry machinery to
snapshots.ts, and the pure ProseMirror sketch operations to
sketch/doc_ops.ts. Define syncSuggesting once instead of twice.
milkdown_editor.tsx drops from 901 to 693 lines with behavior unchanged.
